requirements.txt
streamlit
google-generativeai
chromadb
beautifulsoup4